Institutions feature matrix
Interestingly, we are better with your features than with place of birth. With the following attributes, we achieve the accuracy below.
Variable                                            True
========================================================
URLinText                                         5.2494
numberOfTheSentenceInSnippetWhereSubIsPresent     0.9641
subjInRelation                                    1.8651
=== Stratified cross-validation ===
=== Summary ===
Correctly Classified Instances       27574               75.9112 %
Incorrectly Classified Instances      8750               24.0888 %

The majority baseline is as follows:
Total # of relations: 36324
Positive (majority): 25679
Positive relations in %: 70.69%

I think this is actually quite good.